Tests/LibWeb/Text/expected/shadow-root-adopted-style-sheets.txt

border of #test = (2px solid rgb(173, 255, 47))
